+++ Changelog.md
@@ -3,7 +3,7 @@
 ## NEXT
 
 - Reword message when insufficient ammunition (not just bullets) is available based on number of targets or selected burst.
-- Set **default height** of SK sheet to 768, to match the minimum height expected by Foundry.
+- Ensure height of Stormknight sheet doesn't overflow the available window height.
 - Add icon (with tooltip) to each deck in the **Cards Directory** (right sidebar) indicating those decks which have been configured in the Default Card Deck Configuration window.
 - Allow GM to click in the "Type" column of a Player Hand to **open the deck** from which that card came.
 - Add option for player to **return a card** to its original card (e.g. if the player drew a card from the wrong deck) by holding SHIFT while pressing the "Discard" button.